Array Queuing issues of all flavors.
cards.push (line 48) adds to the cards array, but not the array that
that cards.each (line 39) has already loaded. I think. What confuses me
is that puts cards (line 49) puts the whole, edited array.
Actually, that might be good. The script does retest the when a wrong
guess is given. Then we can, quite simply, write a block of code for
counting frequency.